# Conflicts:
#	engine/unit_definitions.txt

+++ engine/frontend/widgets/panels/stellar_orbit_panel.py
@@ -54,7 +54,7 @@ class OrbitPanel(BaseWidget):
         self.planet_area = AvailablePlanets(self, ANCHO - 200, 32, 200, 340)
         self.recomendation = Recomendation(self, 80, ALTO // 2 - 130)
 
-        self._loaded_orbits = []
+        self._loaded_orbits = {}
         self.indexes = []
 
         self._orbits = {}
@@ -86,9 +86,9 @@ class OrbitPanel(BaseWidget):
         star = Systems.get_current_star()
         self.current = star
         self.curr_idx = self.indexes.index(star)
-        self.orbits = self._orbits[star]
-        self.markers = self._markers[star]
-        self.buttons = self._buttons[star]
+        self.orbits = self._orbits[star.id]
+        self.markers = self._markers[star.id]
+        self.buttons = self._buttons[star.id]
         if not len(self.markers) or not self.markers[0].locked:
             self.populate()
         self.toggle_current_markers_and_buttons(True)
@@ -108,7 +108,7 @@ class OrbitPanel(BaseWidget):
         for marker in markers:
             x = OrbitMarker(self, marker, star, markers[marker])
             x.locked = True
-            self._markers[star].append(x)
+            self._markers[star.id].append(x)
             self.properties.add(x, layer=4)
 
         if hasattr(star, 'habitable_orbit'):
@@ -119,9 +119,8 @@ class OrbitPanel(BaseWidget):
             for marker in markers:
                 x = OrbitMarker(self, marker, star, markers[marker])
                 x.locked = True
-                self._markers[star].append(x)
+                self._markers[star.id].append(x)
                 self.properties.add(x, layer=4)
-            self.add_orbit_marker(star.habitable_orbit)
 
         self.sort_markers()
 
@@ -155,8 +154,8 @@ class OrbitPanel(BaseWidget):
 
         if test is True:
             new = OrbitMarker(self, 'Orbit', star, position, is_orbit=ba, is_complete_orbit=bb, is_resonance=bc)
-            self._markers[star].append(new)
-            self._orbits[star].append(new)
+            self._markers[star.id].append(new)
+            self._orbits[star.id].append(new)
             self.sort_markers()
             self.add_button_and_type(star, new, color)
             self.properties.add(new, layer=4)
@@ -164,7 +163,7 @@ class OrbitPanel(BaseWidget):
     def add_button_and_type(self, star, marker, color):
         orbit_type = OrbitType(self)
         button = OrbitButton(self, color)
-        self._buttons[star].append(button)
+        self._buttons[star.id].append(button)
 
         # Buttons, OrbitTypes and Markers are all Intertwined.
         orbit_type.intertwine(m=marker, b=button)
@@ -213,6 +212,7 @@ class OrbitPanel(BaseWidget):
                 self.area_modify.unlink()
             idx = self.markers.index(marker)
             del self.markers[idx]
+            self._orbits[marker.orbit.star.id].remove(marker)
             self.buttons.remove(marker.linked_button)
             self.sort_markers()
             self.sort_buttons()
@@ -268,20 +268,24 @@ class OrbitPanel(BaseWidget):
             self.clear_ratios()
 
     def save_orbits(self, event):
-        orbits = self._loaded_orbits
         for system in Systems.get_systems():
             if system.star_system.letter == 'S':
                 for star in system:
-                    for marker in self._orbits.get(star, []):
+                    for marker in self._orbits.get(star.id, []):
+                        astrobody_id = marker.orbit.astrobody.id
                         d = self.create_save_data(marker.orbit)
-                        orbits.append(d)
+                        self._loaded_orbits[astrobody_id] = d
             else:
                 star = system.star_system
-                for marker in self._orbits.get(star, []):
+                for marker in self._orbits.get(star.id, []):
                     d = self.create_save_data(marker.orbit)
-                    orbits.append(d)
+                    if not isinstance(marker.orbit, (RawOrbit, PseudoOrbit)):
+                        astrobody_id = marker.orbit.astrobody.id
+                        if astrobody_id not in self._loaded_orbits:
+                            self._loaded_orbits[astrobody_id] = d
 
-        EventHandler.trigger(event.tipo + 'Data', 'Orbit', {'Stellar Orbits': orbits})
+        EventHandler.trigger(event.tipo + 'Data', 'Orbit', {'Stellar Orbits': self._loaded_orbits})
+        self._loaded_orbits.clear()
 
     @staticmethod
     def create_save_data(orb):
@@ -298,24 +302,29 @@ class OrbitPanel(BaseWidget):
         return d
 
     def load_orbits(self, event):
-        for position in event.data.get('Stellar Orbits', []):
-            if position not in self._loaded_orbits:
-                self._loaded_orbits.append(position)
+        for id in event.data.get('Stellar Orbits', []):
+            position = event.data['Stellar Orbits'][id]
+            if id not in self._loaded_orbits:
+                self._loaded_orbits[id] = position
 
     def set_loaded_orbits(self):
-        for orbit_data in self._loaded_orbits:
-            a = q(orbit_data['a'], 'au')
-            if 'e' not in orbit_data:
-                self.add_orbit_marker(a)
-            else:
-                e = q(orbit_data['e'])
-                i = q(orbit_data['i'], 'degree')
-                system = Systems.get_system_by_id(orbit_data['star_id'])
-                planet = system.get_astrobody_by(orbit_data['astrobody'], tag_type='id')
-                star = system.star_system
-                planet.set_orbit(star, [a, e, i])
-                self.add_orbit_marker(planet.orbit)
-                self.planet_area.delete_objects(planet)
+        existing_ids = [marker.orbit.id for marker in self.orbits]
+        for id in self._loaded_orbits:
+            if id not in existing_ids:
+                orbit_data = self._loaded_orbits[id]
+                a = q(orbit_data['a'], 'au')
+                if 'e' not in orbit_data:
+                    self.add_orbit_marker(a)
+                else:
+                    e = q(orbit_data['e'])
+                    i = q(orbit_data['i'], 'degree')
+                    system = Systems.get_system_by_id(orbit_data['star_id'])
+                    planet = system.get_astrobody_by(id, tag_type='id')
+                    star = system.star_system
+                    planet.set_orbit(star, [a, e, i])
+                    planet.orbit.id = id
+                    self.add_orbit_marker(planet.orbit)
+                    self.planet_area.delete_objects(planet)
 
         # borrar las órbitas cargadas para evitar que se dupliquen.
         self.sort_markers()
@@ -325,13 +334,14 @@ class OrbitPanel(BaseWidget):
         assert len(Systems.get_systems())
         for system in Systems.get_systems():
             star = system.star_system
-            if star not in self._markers:
-                self._markers[star] = []
-                self._orbits[star] = []
-                self._buttons[star] = []
+            if star.id not in self._markers:
+                self._markers[star.id] = []
+                self._orbits[star.id] = []
+                self._buttons[star.id] = []
                 self.indexes.append(star)
 
     def show(self):
+        super().show()
         try:
             self.fill_indexes()
             self.set_current()
@@ -346,8 +356,6 @@ class OrbitPanel(BaseWidget):
         if len(self._loaded_orbits):
             self.set_loaded_orbits()
         self.show_markers_button.show()
-
-        super().show()
 
     def hide(self):
         super().hide()
@@ -436,7 +444,7 @@ class OrbitPanel(BaseWidget):
         x = int(self.digit_x.value)
         y = int(self.digit_y.value)
         assert x >= y, 'invalid ratio'
-        self.write('{}° Order'.format(x-y), self.order_f, right=self.digit_x.rect.left - 2, y=self.digit_x.rect.y)
+        self.write('{}° Order'.format(x - y), self.order_f, right=self.digit_x.rect.left - 2, y=self.digit_x.rect.y)
         return '{}:{}'.format(x, y)
 
     def clear_ratios(self):
@@ -542,8 +550,9 @@ class OrbitType(BaseWidget, Intertwined):
                 value = 'au'
             parametros.append(value)
         main = self.parent.current
-        orbit = self.linked_astrobody.set_orbit(main, parametros)
-        self.linked_marker.orbit = orbit
+        if self.linked_astrobody.orbit is None:
+            orbit = self.linked_astrobody.set_orbit(main, parametros)
+            self.linked_marker.orbit = orbit
         self.show()
         self.parent.planet_area.delete_objects(self.linked_astrobody)
         if hasattr(self.parent, 'recomendation'):
@@ -865,6 +874,10 @@ class Recomendation(BaseWidget):
         e = round(0.584 * pow(planets_in_system, -1.2), 3) if planets_in_system > 1 else None
         if planet.habitable and orbit.temperature == 'habitable':
             data = recomendation['habitable']
+            if hasattr(star, 'habitable_orbit'):
+                # though this may never come into play.
+                if orbit.a.m <= star.habitable_orbit:
+                    data = recomendation['inner']
             if e is not None and e <= 0.2:
                 data.update({'e': e})
